# api/src/docs/swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/invoice/write":{"post":{"operationId":"InvoiceController_add","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvoiceDTO"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["invoice"]}},"/invoice/{id}/generate":{"post":{"operationId":"InvoiceController_generate","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["invoice"]}},"/client/add":{"post":{"operationId":"ClientController_add","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["client"]}},"/client/list":{"get":{"operationId":"ClientController_findOne","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["client"]}},"/client/{id}":{"patch":{"operationId":"ClientController_updateClient","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["client"]},"delete":{"operationId":"ClientController_deleteClient","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["client"]}},"/timesheet/{clientId}/add":{"post":{"operationId":"TimeSheetController_add","parameters":[{"name":"clientId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TimeSheetDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["timesheet"]}},"/timesheet/{clientId}/list":{"get":{"operationId":"TimeSheetController_findOne","parameters":[{"name":"clientId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["timesheet"]}},"/timesheet/{id}":{"patch":{"operationId":"TimeSheetController_updateClient","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TimeSheetDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["timesheet"]},"delete":{"operationId":"TimeSheetController_deleteClient","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["timesheet"]}},"/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["users"]}},"/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["users"]}},"/auth/login":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"]}}},"info":{"title":"Bill Tracker API","description":"Web service for invoices","version":"1.0","contact":{}},"tags":[{"name":"users","description":"User details management"},{"name":"auth","description":"Authentication"}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"WorkItemDTO":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}},"hours":{"type":"number"}},"required":["title","description","tags","hours"]},"InvoiceDTO":{"type":"object","properties":{"clientId":{"type":"number"},"date":{"type":"string","format":"date-time"},"note":{"type":"string"},"workItems":{"type":"array","items":{"$ref":"#/components/schemas/WorkItemDTO"}}},"required":["clientId","date","note","workItems"]},"ClientDto":{"type":"object","properties":{"name":{"type":"string"},"code":{"type":"string"},"symbol":{"type":"string"},"address":{"type":"string"},"hourly_rate":{"type":"number"},"hours_per_day":{"type":"number"},"category":{"type":"string"}},"required":["name","code","symbol","address","hourly_rate","hours_per_day","category"]},"TimeSheetDto":{"type":"object","properties":{"summary":{"type":"string"},"entry_date":{"format":"date-time","type":"string"},"tags":{"type":"array","items":{"type":"string"}}},"required":["summary","entry_date","tags"]},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string","default":""},"password":{"type":"string","default":""},"name":{"type":"string","default":""},"address":{"type":"string","default":""},"bank_name":{"type":"string","default":""},"bank_swift_code":{"type":"string","default":""},"bank_account_number":{"type":"string","default":""},"bank_account_name":{"type":"string","default":""}},"required":["email","password","name","address","bank_name","bank_swift_code","bank_account_number","bank_account_name"]},"SignInDto":{"type":"object","properties":{"email":{"type":"string","default":""},"password":{"type":"string","default":""}}}}}}